+++ modules/spring-core/src/main/kotlin/com/esprito/spring/core/providers/ConfigurationPropertyLineMarkerProvider.kt
@@ -3,28 +3,23 @@ package com.esprito.spring.core.providers
 import com.esprito.spring.core.SpringCoreBundle
 import com.esprito.spring.core.SpringIcons
 import com.esprito.spring.core.SpringProperties.COLON
-import com.esprito.spring.core.properties.ConfigurationPropertyDataRetriever
+import com.esprito.spring.core.properties.dataRetriever.ConfigurationPropertyDataRetrieverFactory
 import com.intellij.codeInsight.daemon.RelatedItemLineMarkerInfo
 import com.intellij.codeInsight.daemon.RelatedItemLineMarkerProvider
 import com.intellij.codeInsight.navigation.NavigationGutterIconBuilder
 import com.intellij.codeInsight.navigation.impl.PsiTargetPresentationRenderer
 import com.intellij.json.psi.JsonStringLiteral
-import com.intellij.lang.Language
 import com.intellij.lang.properties.psi.Property
 import com.intellij.openapi.editor.markup.GutterIconRenderer
 import com.intellij.openapi.module.ModuleUtilCore
 import com.intellij.openapi.util.NotNullLazyValue
-import com.intellij.openapi.util.text.StringUtil
 import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiField
 import com.intellij.psi.presentation.java.SymbolPresentationUtil
-import com.intellij.psi.util.childrenOfType
-import org.jetbrains.kotlin.idea.KotlinLanguage
-import org.jetbrains.kotlin.psi.KtPropertyAccessor
-import org.jetbrains.uast.*
+import org.jetbrains.uast.getUParentForIdentifier
 import org.jetbrains.yaml.YAMLUtil
 import org.jetbrains.yaml.psi.YAMLKeyValue
 import javax.swing.Icon
+import com.esprito.spring.core.properties.dataRetriever.ConfigurationPropertyDataRetriever.Companion as DataRetriever
 
 class ConfigurationPropertyLineMarkerProvider : RelatedItemLineMarkerProvider() {
 
@@ -33,19 +28,13 @@ class ConfigurationPropertyLineMarkerProvider : RelatedItemLineMarkerProvider() 
         result: MutableCollection<in RelatedItemLineMarkerInfo<*>>
     ) {
         val uParent = getUParentForIdentifier(element) ?: return
-        val uMethod = when (uParent) {
-            is UMethod -> uParent
-            is UField -> getMethodFromKtField(uParent)
-            else -> null
-        } ?: return
-
-        val dataRetriever = ConfigurationPropertyDataRetriever(uMethod.javaPsi)
+        val dataRetriever = ConfigurationPropertyDataRetrieverFactory.createFor(uParent) ?: return
         val psiClass = dataRetriever.getContainingClass() ?: return
         val memberName = dataRetriever.getMemberName() ?: return
         val module = ModuleUtilCore.findModuleForPsiElement(element) ?: return
-        val nameElement = uMethod.uastAnchor?.sourcePsi ?: return
+        val nameElement = dataRetriever.getNameElementPsi() ?: return
 
-        val prefixValue = ConfigurationPropertyDataRetriever.getPrefixValue(psiClass, module)
+        val prefixValue = DataRetriever.getPrefixValue(psiClass, module)
         if (prefixValue.isBlank()) return
 
         val properties = dataRetriever.getRelatedProperties(prefixValue, memberName, module)
@@ -63,23 +52,6 @@ class ConfigurationPropertyLineMarkerProvider : RelatedItemLineMarkerProvider() 
             .setTargetRenderer { getTargetRender() }
 
         result += builder.createLineMarkerInfo(nameElement)
-    }
-
-    private fun getMethodFromKtField(uField: UField): UMethod? {
-        val psiField = uField.javaPsi as? PsiField ?: return null
-        if (psiField.language != KotlinLanguage.INSTANCE) return null
-
-        val name = "set${StringUtil.capitalize(psiField.name)}"
-
-        val setter = psiField.toUElement()
-            ?.sourcePsi
-            ?.childrenOfType<KtPropertyAccessor>()
-            ?.firstOrNull { it.isSetter }
-
-        return if (setter == null) {
-            uField.getUastParentOfType<UClass>()?.methods
-                ?.firstOrNull { it.name == name }
-        } else null
     }
 
     private fun getTargetRender(): PsiTargetPresentationRenderer<PsiElement> {